requirements.txt
streamlit
python-dotenv
openai
pypdf
requests
langchain
chromadb
